refactor setup script
