+++ public/themes/default/pages/admin.html
@@ -37,11 +37,6 @@
       include
       "./fragments/admin/servers.html"
     %}
-
-    <footer>
-      <a href="https://github.com/Pomax/make-webbly-things/issues"
-        >File issues here</a
-      >
-    </footer>
+    {% include "fragments/file-issues.html" %}
   </body>
 </html>
