+++ public/themes/default/pages/fragments/file-issues.html
@@ -1,0 +1,5 @@
+<footer>
+  <a href="https://github.com/Pomax/make-webbly-things/issues"
+    >File platform issues here</a
+  >
+</footer>
