+++ public/themes/default/pages/main/logged-in.html
@@ -28,7 +28,7 @@
     <button disabled>Let's go!</button>
   </div>
 
-  {% if projectList %}
+  {% if projectList.length %}
     <h2>Your projects:</h2>
 
     <ul>
@@ -43,9 +43,4 @@
   {% endif %}
 {% endif %}
 {% include "fragments/settings.html" %}
-
-<footer>
-  <a href="https://github.com/Pomax/make-webbly-things/issues"
-    >File issues here</a
-  >
-</footer>
+{% include "../fragments/file-issues.html" %}
